Add possibility to create nuget package
