+++ samples/GStreamerWPF/GStreamerControl.Demo.Direct/MainWindow.xaml
@@ -1,0 +1,26 @@
+﻿<Window x:Class="GStreamerD3DSampleCore.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:component="clr-namespace:GStreamerControl.Library;assembly=GStreamerControl.Library"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+    <Grid x:Name="grid" >
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="*"/>
+        </Grid.ColumnDefinitions>
+        <component:GStreamerView x:Name="stream1" Grid.Column="0" Rtsp="True" Enabled="True"/>
+        <component:GStreamerView x:Name="stream2" Grid.Column="1" Rtsp="False" Enabled="True"/>
+        <TextBox Text="Overlay test"                 
+                 VerticalAlignment="Center"
+                 HorizontalContentAlignment="Center"
+                 VerticalContentAlignment="Center"
+                 Background="#44000000"
+                 FontSize="30" 
+                 Grid.ColumnSpan="2"/>
+        <Button Click="Button_Click" Grid.Column="2"/>
+    </Grid>
+</Window>
