+++ samples/GStreamerWPF/GStreamerControl.Demo.Direct/log4net.config
@@ -1,0 +1,23 @@
+﻿<log4net>
+  <root>
+    <level value="ALL" />
+    <appender-ref ref="console" />
+    <appender-ref ref="file" />
+  </root>
+  <appender name="console" type="log4net.Appender.ConsoleAppender">
+    <layout type="log4net.Layout.PatternLayout">
+      <conversionPattern value="%date %level %logger - %message%newline" />
+    </layout>
+  </appender>
+  <appender name="file" type="log4net.Appender.RollingFileAppender">
+    <file value="logfile.txt" />
+    <appendToFile value="true" />
+    <rollingStyle value="Size" />
+    <maxSizeRollBackups value="5" />
+    <maximumFileSize value="10MB" />
+    <staticLogFileName value="true" />
+    <layout type="log4net.Layout.PatternLayout">
+      <conversionPattern value="%date [%thread] %level %logger - %message%newline" />
+    </layout>
+  </appender>
+</log4net>
