+++ samples/GStreamerWPF/GStreamerControl.Demo.Nuget/GStreamerControl.Demo.Nuget.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net6.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
+    <UseWPF>true</UseWPF>
+	  <PlatformTarget>x64</PlatformTarget>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="GStreamer.WPF" Version="1.0.2" />
+    <PackageReference Include="GstSharp" Version="1.20.2.1" />
+  </ItemGroup>
+
+</Project>
