+++ samples/GStreamerWPF/GStreamerControl.Demo.Nuget/MainWindow.xaml
@@ -1,0 +1,13 @@
+﻿<Window x:Class="finaltest.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:wpfstream="clr-namespace:GStreamerControl.Library;assembly=GStreamerControl.Library"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="450" Width="800">
+    <Grid>
+        <wpfstream:GStreamerView x:Name="stream1" Rtsp="False" Enabled="True"></wpfstream:GStreamerView>
+        <Button Content="salut" HorizontalAlignment="Left" VerticalAlignment="Top"></Button>
+    </Grid>
+</Window>
