+++ samples/GStreamerWPF/GStreamerControl.Demo.Nuget/MainWindow.xaml.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Windows;
+
+namespace finaltest
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+        }
+
+        protected override void OnClosed(EventArgs e)
+        {
+            base.OnClosed(e);
+            stream1.Dispose();
+        }
+    }
+}
