--- samples/GStreamerWPF/GStreamerD3DSampleCore/App.xaml
@@ -1,9 +1,0 @@
-﻿<Application x:Class="GStreamerD3DSampleCore.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:GStreamerD3DSampleCore"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
