--- samples/GStreamerWPF/WPFStream/StreamComponent.xaml
@@ -1,12 +1,0 @@
-﻿<UserControl x:Class="WPFStream.StreamComponent"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:WPFStream"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Grid>
-        <Image x:Name="d3dScene" />
-    </Grid>
-</UserControl>
